Solutions for Day 12
Despite an infuriating lack of respect for physics terminology (no, you can not multiply PE and KE to get total Energy), part 1 wasn't bad. Part 2 however was pretty tricky to come up with an algorithm for. Eventually I decided to check Reddit for hints. The answer was really simple and obvious in retrospect...

I probably should have kept thinking on the problem before looking for hints but ¯\_(ツ)_/¯