Solutions for day 16.
This day was _hard_. Part 1 was fairly simple but part 2 required some out of the box thinking (thanks for the hints Reddit) that makes my program not work for the general case.
Note to self: next time an AoC challenge looks like it will require lots of computation figure out if it can be solved as a specific case of the general problem.